+++ app/views/layout/navbar.scala.html
@@ -17,7 +17,7 @@
             <nav>
                 <ul class="nav navbar-nav">
                     <li class="active">
-                        <a href="@routes.ProjectController.getProjectWidgets(0)">Projekte</a>
+                        <a href="@routes.ApplicationController.index()">Projekte</a>
                     </li>
                     @if(session.get("email") != null){
                         <li>
