+++ app/views/project/list/projectList.scala.html
@@ -20,15 +20,15 @@
                             @if(pageSize > 1) {
                                 @if(currentPage > 1) {
                                     <li>
-                                        <a href="@routes.ProjectController.getProjectWidgets(currentPage - 1)">Zurück</a>
+                                        <a href="@routes.ProjectListController.getProjectWidgets(currentPage - 1)">Zurück</a>
                                     </li>
                                 }
                                 @for(i <- 0 to pageSize - 1) {
-                                    <li @if(currentPage == i){class="active"}><a href="@routes.ProjectController.getProjectWidgets(i)">@i</a></li>
+                                    <li @if(currentPage == i){class="active"}><a href="@routes.ProjectListController.getProjectWidgets(i)">@i</a></li>
                                 }
                                 @if(currentPage < pageSize) {
                                     <li class="active">
-                                        <a href="@routes.ProjectController.getProjectWidgets(currentPage + 1)">Weiter</a>
+                                        <a href="@routes.ProjectListController.getProjectWidgets(currentPage + 1)">Weiter</a>
                                     </li>
                                 }
                             } else {
