+++ app/views/project/list/projectListMain.scala.html
@@ -1,0 +1,5 @@
+@(widgets: List[viewmodels.ProjectWidget], pageSize: Int, currentPage: Int)
+
+@main("BringToAfrica") {
+       @projectList(widgets,pageSize,currentPage)
+} { }
