+++ app/views/startpage.scala.html
@@ -8,7 +8,7 @@
                     </header>
                 </div>
                 <div>
-                    @ProjectController.getProjectWidgetList(0)
+                    @ProjectListController.getProjectWidgetList(0)
                 </div>
             </div>
