+++ app/views/user/myDonations.scala.html
@@ -14,7 +14,7 @@
     <div class="row">
         <div class="col-xs-12">
             <h4>Du hast bis jetzt noch keine Spende getätigt.
-                <a href="@routes.ProjectController.getProjectWidgets(0)">Jetzt zu den Projekten</a>
+                <a href="@routes.ProjectListController.getProjectWidgets(0)">Jetzt zu den Projekten</a>
             </h4>
         </div>
     </div>
